requirements.txt
flask==2.1.2
werkzeug==2.1.2
langchain==0.0.147
openai==0.27.6
python-dotenv==1.0.0
flask-cors==4.0.1
